prep for custom email senders: moving email sending to a separate file
